use new integers path

+++ compiler/src/field/mod.rs
@@ -14,7 +14,7 @@ use snarkos_models::{
             boolean::Boolean,
             eq::{ConditionalEqGadget, EqGadget},
             select::CondSelectGadget,
-            uint8::UInt8,
+            uint::UInt8,
             ToBitsGadget, ToBytesGadget,
         },
     },
